🐛 Pydantic V1 会把 float 转换成 int

+++ nonebot_plugin_tetris_stats/utils/typing.py
@@ -1,7 +1,7 @@
 from collections.abc import Awaitable, Callable
 from typing import Any, Literal
 
-Number = int | float
+Number = float | int
 GameType = Literal['IO', 'TOP', 'TOS']
 CommandType = Literal['bind', 'query']
 AsyncCallable = Callable[..., Awaitable[Any]]
